src/aimm/requirements.txt
pillow
Flask
Flask-WTF
torch
torchvision
scikit-learn
pandas
tensorflow
python-dotenv
numpy
redis
redisearch